requirements.txt
cycler==0.10.*
cython==0.29.*
matplotlib==3.0.*
numpy==1.17.* # for cocoeval
opencv-python==3.4.*
Pillow==5.4.*
prefetch-generator
# pycocotools==2.0
pyparsing==2.4.*
scikit-image
six
torch==1.3.*
torchsummary
torchvision==0.2.*
tqdm
onnx==1.5.0
onnx-coreml==1.3
onnxruntime==1.3.0
coremltools==3.4
